move SceneSpatialPrimitives to its own mod and picking along with it
